[NCL-7076] - remove old BPM invocation code - improve.
