--- bpm/src/main/java/org/jboss/pnc/bpm/MockConnector.java
@@ -1,57 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2022 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.bpm;
-
-import org.jboss.pnc.spi.exception.ProcessManagerException;
-
-/**
- * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
- */
-public class MockConnector implements Connector {
-    @Override
-    public Long startProcess(String processId, Object processParameters, String accessToken)
-            throws ProcessManagerException {
-        return 1L;
-    }
-
-    @Override
-    public Long startProcess(String processId, Object requestObject, String correlationKey, String accessToken)
-            throws ProcessManagerException {
-        return 1L;
-    }
-
-    @Override
-    public boolean isProcessInstanceCompleted(Long processInstanceId) {
-        return false;
-    }
-
-    @Override
-    public boolean cancelByCorrelation(String correlationKey, String accessToken) {
-        return false;
-    }
-
-    @Override
-    public boolean cancel(Long processInstanceId, String accessToken) {
-        return false;
-    }
-
-    @Override
-    public void close() {
-
-    }
-}
