+++ bpm/src/main/java/org/jboss/pnc/bpm/model/ComponentBuildParameters.java
@@ -49,16 +49,6 @@ public class ComponentBuildParameters implements Serializable {
     @Getter
     private String daBaseUrl;
 
-    @Deprecated // not used anymore
-    @NotNull
-    @Getter
-    private Boolean communityBuild = false;
-
-    @Deprecated // not used anymore
-    @NotNull
-    @Getter
-    private Boolean versionAdjust = false;
-
     @NotNull
     @Getter
     BuildExecutionConfigurationRest buildExecutionConfiguration;
