+++ bpm/src/main/java/org/jboss/pnc/bpm/task/BpmBuildTask.java
@@ -55,8 +55,6 @@ public class BpmBuildTask extends BpmTask {
                 globalConfig.getExternalIndyUrl(),
                 globalConfig.getExternalRepourUrl(),
                 globalConfig.getExternalDaUrl(),
-                false,
-                false,
                 getBuildExecutionConfiguration(buildTask));
     }
